requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
coloredlogs==15.0.1
discord.py==2.5.2
flatbuffers==25.2.10
frozenlist==1.6.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
multidict==6.4.3
nudenet==3.4.2
numpy==2.2.5
onnxruntime==1.21.1
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
pyreadline3==3.5.4
python-dotenv==1.1.0
requests==2.32.3
sympy==1.14.0
urllib3==2.4.0
yarl==1.20.0
